requirement.txt
streamlit
google-generativeai
python
PIl